chart-svc/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
pandas
mplfinance
requests
